Las bolas tienen un bug cuando hay colisiones simultaneas

+++ app/src/main/java/self/joanciscar/gamedevelopment/GameEntity.java
@@ -20,8 +20,8 @@ public interface GameEntity {
 
     boolean hasPosition();
 
-    void processDump(GameEntity gameEntity, double height, double width, double min_height, double min_width,boolean callOther);
-    default void processDump(GameEntity gameEntity, double height, double width, double min_height, double min_width) {
+    void processDump(GameEntity gameEntity, float height, float width, float min_height, float min_width,boolean callOther);
+    default void processDump(GameEntity gameEntity, float height, float width, float min_height, float min_width) {
         this.processDump(gameEntity,height,width,min_height,min_width,true);
     }
     void setPosition(Vector position);
